# data-collector/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.NewsListResponse:
    properties:
      limit:
        description: 每页数量
        type: integer
      list:
        description: 新闻列表
        items:
          $ref: '#/definitions/models.News'
        type: array
      offset:
        description: 偏移量
        type: integer
      total:
        description: 总数
        type: integer
    type: object
...
  handler.Response:
    properties:
      code:
        description: 响应码
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
        type: string
      message:
        description: 响应消息
        type: string
    type: object
  models.News:
    properties:
      content:
        description: 快讯内容
        type: string
      created_at:
        description: 创建时间
        type: string
      id:
        type: string
      publish_time:
        description: 发布时间
        type: string
      related_industries:
        description: 关联行业
        items:
          type: string
        type: array
      related_stocks:
        description: 关联股票
        items:
          $ref: '#/definitions/models.RelatedStock'
        type: array
      source:
        description: 来源（如：财联社）
        type: string
      title:
        description: 快讯标题
        type: string
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 原文链接
        type: string
    type: object
...
  models.RelatedStock:
    properties:
      code:
        description: 股票代码
        type: string
      name:
        description: 股票名称
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 数据采集服务API文档，提供股票、财务、新闻和政策数据采集功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Data Collector API
  version: "1.0"
paths:
  /api/v1/news:
    get:
      consumes:
      - application/json
      description: 根据条件获取新闻列表，支持分页
      parameters:
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      - description: 新闻来源
        in: query
        name: source
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - description: 关联股票代码
        in: query
        name: stock_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.NewsListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取新闻列表
      tags:
      - 新闻管理
  /api/v1/news/by-stock:
    get:
      consumes:
      - application/json
      description: 获取与指定股票相关的新闻列表
      parameters:
      - description: 股票代码
        in: path
        name: stock_code
        required: true
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.NewsListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据股票代码获取相关新闻
      tags:
      - 新闻管理
  /api/v1/news/detail:
    get:
      consumes:
      - application/json
      description: 根据新闻ID获取详细信息
      parameters:
      - description: 新闻ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.News'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取新闻详情
      tags:
      - 新闻管理
  /api/v1/news/search:
    get:
      consumes:
      - application/json
      description: 根据关键词搜索新闻
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.NewsListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 搜索新闻
      tags:
      - 新闻管理
  /api/v1/news/time-range:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的新闻列表
      parameters:
      - description: 开始时间
        format: date-time
        in: query
        name: start_time
        required: true
        type: string
      - description: 结束时间
        format: date-time
        in: query
        name: end_time
        required: true
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.NewsListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据时间范围获取新闻
      tags:
      - 新闻管理
...
schemes:
- http
- https
swagger: "2.0"